feat(components): separate class and template exports
