+++ src/components/test-input/test-input.js
@@ -8,45 +8,14 @@
  */
 
 import settings from '../../globals/settings.js';
-import { LitElement, html } from 'lit';
-import { SearchTypeaheadAPI } from '../../services/SearchTypeahead/index.js';
+import testInput from './src/test-input.js';
+import { testInputTemplate } from './src/test-input.template.js';
 
 const { stablePrefix: c4aiPrefix } = settings;
 
-/**
- * Input component using search typeahead api
- */
-export default class C4AITestInput extends LitElement {
-  static properties = {
-    searchResults: { attribute: false },
-  };
-
-  constructor() {
-    super();
-    this.searchResults;
-  }
-
-  async getResults(searchQuery) {
-    const response = await SearchTypeaheadAPI.getResults(searchQuery);
-    return response.map((res) => res[0]);
-  }
-
-  _handleInput(event) {
-    const { value } = event.target;
-    this.getResults(value).then((res) => {
-      this.searchResults = res;
-    });
-  }
-
+class C4AITestInput extends testInput {
   render() {
-    const { _handleInput: handleInput } = this;
-    return html`<div>
-      <label>Search typeahead</label>
-      <input type="text" @input="${handleInput}" />
-      ${this.searchResults
-        ? this.searchResults.map((result) => html`<p>${result}</p>`)
-        : undefined}
-    </div>`;
+    return testInputTemplate(this);
   }
 }
